fix(y4m): add (size_t) cast in y4m_convert_411_422jpeg pointer advance
`_dst += _y4m->pic_w * _y4m->pic_h` used plain int * int arithmetic
before advancing an unsigned char pointer. For frames where
pic_w * pic_h exceeds INT_MAX (~46340 x 46340), this is signed-integer
overflow — undefined behaviour under C99/C11 (SEI CERT C INT30-C /
INT32-C). Fix: `_dst += (size_t)_y4m->pic_w * _y4m->pic_h`, matching
the identical cast already present in the sibling functions
y4m_convert_42xmpeg2_42xjpeg (line 220) and
y4m_convert_42xpaldv_42xjpeg (line 315). Added a matching explanatory
comment to keep the three functions consistent. docs/state.md updated
per CLAUDE §12 r13.

+++ core/tools/y4m_input.c
@@ -491,8 +491,10 @@ static void y4m_convert_411_422jpeg(y4m_input *_y4m, unsigned char *_dst, unsign
     int pli;
     int y;
     int x;
-    /*Skip past the luma data.*/
-    _dst += _y4m->pic_w * _y4m->pic_h;
+    /*Skip past the luma data.
+     * Cast to size_t to avoid signed-integer overflow for large frames
+     * (pic_w * pic_h overflows int when each dimension exceeds ~46340). */
+    _dst += (size_t)_y4m->pic_w * _y4m->pic_h;
     /*Compute the size of each chroma plane.*/
     c_w = (_y4m->pic_w + _y4m->src_c_dec_h - 1) / _y4m->src_c_dec_h;
     dst_c_w = (_y4m->pic_w + _y4m->dst_c_dec_h - 1) / _y4m->dst_c_dec_h;
